Don't move the cursor to beginning of line on deactivating limelight
Calling functions in vim using a range with `<line1>,<line2>call ...`
makes vim move the cursor to the beginning of the line. This can cause
distraction when Limelight is used in conjunction with the insert mode
(e.g. when deactivating on `InsertLeave`).

This commit makes the `Limelight` command pass the range as parameters
to the `execute` function which removes that effect while
`Limelight`ening a range still works.